- Created first tests  - Added steps to 1) launch browser. 2) navigate to Google, 3) Enter search terms 4) Click Search button, getting errors, cleaned some formatting 5) fixed an error that came from trying to click a Search button. Error was coming up because I needed to accept a EULA for something in intellij.. maybe? clicked to accept and the error stopped. 6) added implicit wait period of 5 seconds to search step to avoid timing issue/updated search terms 7) edited line 63 to search for a partial link name, that let selenium find the desired link and click. Added a short sleep timer to let me see the successful loading of the desired page before closing the browser
2019-06-05
1) Added steps to log into State Portal UI,
2) kinda stuck trying to add an Assert
3) Added steps to navigate to automationPractice.com, click a dress, add a quantity
4) Fixed comments

2019-06-06
1) trying to add assert statement to check the math on the sample website. struggling to figure that out, fixed variable type to allow decimals

+++ src/test/java/com/tim/GooglePageTests.java
@@ -140,8 +140,8 @@ public class GooglePageTests {
         printedDressQuantity.sendKeys("999");
         WebElement addToCartButton = driver.findElement(name("Submit"));
         addToCartButton.click();
-        long expected;
-        long actual;
+        float expected;
+        float actual;
         Assert.assertEquals((actual = "25,974.00", expected = (999*26));
